+++ src/lib/fn.js
@@ -97,3 +97,10 @@ export const useDirective = binding => {
     }
   });
 };
+
+export const useVisibleViewport = el => {
+  const { right } = el.getBoundingClientRect();
+  const vWidth = window.innerWidth || document.documentElement.clientWidth;
+
+  return right < vWidth;
+};