+++ src/locale/en.js
@@ -1,0 +1,13 @@
+// English [en]
+
+export default {
+  today: 'Today',
+  yesterday: 'Yesterday',
+  past: period => `Last ${period} Days`,
+  currentMonth: 'This Month',
+  pastMonth: 'Last Month',
+  button: {
+    apply: 'Apply',
+    cancel: 'Cancel'
+  }
+};