+++ src/locale/id.js
@@ -1,0 +1,13 @@
+// Indonesia [id]
+
+export default {
+  today: 'Hari ini',
+  yesterday: 'Kemarin',
+  past: period => `${period} hari terakhir`,
+  currentMonth: 'Bulan ini',
+  pastMonth: 'Bulan lalu',
+  button: {
+    apply: 'Terapkan',
+    cancel: 'Batal'
+  }
+};