+++ CHANGELOG.md
@@ -38,7 +38,14 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.7]
 
-- Fix some style
+- Fix some style.
+
+## [1.0.8]
+
+- Add localization for `shortcuts`, `footer` & `dayjs` locale.
+- New props `i18n`, [Example](https://litepie.com/#i18n).
+- Change local model `value` to `pickerValue`.
+- Fix issues placement arrow, `left` to `right`.
 
 [Released]: https://github.com/kenhyuwa/litepie-datepicker/
 [1.0.0]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.0
@@ -49,3 +56,4 @@ All notable changes to this project will be documented in this file.
 [1.0.5]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.5
 [1.0.6]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.6
 [1.0.7]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.7
+[1.0.8]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.8